refactor(ICache): move ICacheReplacer into separate file
